+++ example/searchlight_highlight_example.dart
@@ -1,6 +1,12 @@
 import 'package:searchlight_highlight/searchlight_highlight.dart';
 
 void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
+  final highlighted = Highlight().highlight(
+    'The quick brown fox jumps over the lazy dog',
+    'brown fox',
+  );
+
+  print(highlighted.positions);
+  print(highlighted.HTML);
+  print(highlighted.trim(18));
 }